First pass at moving wib2 tpg code to src/ TPEmulatorModel, WIB@FrameProcessor, RegisterChannelToNumber
